+++ garglk/launchgtk.c
@@ -150,7 +150,7 @@ void winbrowsefile(char *buffer)
 void winpath(char *buffer)
 {
     char exepath[MaxBuffer] = {0};
-    unsigned int exelen;
+    ssize_t exelen;
 
     exelen = readlink("/proc/self/exe", exepath, sizeof(exepath));
